refactor: Meeting 도메인 참조를 app 패키지로 이동시킨다
