+++ src/main/java/net/teumteum/alert/domain/AlertPublisher.java
@@ -1,5 +1,7 @@
 package net.teumteum.alert.domain;
 
+import net.teumteum.alert.domain.Alert;
+
 @FunctionalInterface
 public interface AlertPublisher {
