+++ src/main/java/net/teumteum/alert/domain/AlertService.java
@@ -1,11 +1,8 @@
 package net.teumteum.alert.domain;
 
-import static net.teumteum.alert.domain.AlertExecutorConfigurer.ALERT_EXECUTOR;
-
+import java.util.List;
+import java.util.Set;
 import lombok.RequiredArgsConstructor;
-import net.teumteum.meeting.domain.MeetingAlerted;
-import org.springframework.context.event.EventListener;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -14,13 +11,9 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class AlertService {
 
-    private final AlertPublisher alertPublisher;
     private final AlertRepository alertRepository;
 
-    @Async(ALERT_EXECUTOR)
-    @EventListener({MeetingAlerted.class})
-    public void alert(MeetingAlerted alerted) {
-        alertRepository.findAllByUserId(alerted.userIds())
-            .forEach(alertPublisher::publish);
+    public List<Alert> findAllByUserId(Set<Long> userIds) {
+        return alertRepository.findAllByUserId(userIds);
     }
 }
